requirements.txt
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
distro==1.6.0
Django==3.2.7
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonschema==3.2.0
MarkupSafe==2.0.1
packaging==21.0
paramiko==2.8.0
pycparser==2.20
PyJWT==2.3.0
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dotenv==0.19.1
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
swagger-spec-validator==2.7.3
texttable==1.6.4
uritemplate==4.1.1
urllib3==1.26.7
websocket-client==0.59.0
